reconfigured nextbook using modules
